Rename sse classes and change path
